requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
pytz==2023.3
